==> Notebooks/02_data_cleaning.ipynb <==
import pandas as pd
import numpy as np


spot = pd.read_csv(
    "data/nifty_spot_5min.csv",
    parse_dates=['date'],
    index_col='date'
)

futures = pd.read_csv(
    "data/nifty_futures_5min.csv",
    parse_dates=['date'],
    index_col='date'
)

options = pd.read_csv(
    "data/nifty_options_5min.csv",
    parse_dates=['date'],
    index_col='date'
)


spot.shape, futures.shape, options.shape


#Check for missing timestamps

spot.index.equals(futures.index)


options.index.nunique() == spot.shape[0]


spot.isna().sum()



futures.isna().sum()


options.isna().sum()

spot = spot.sort_index()
futures = futures.sort_index()
options = options.sort_index()

spot = spot[~spot.index.duplicated()]
futures = futures[~futures.index.duplicated()]


options = options.reset_index().drop_duplicates().set_index('date')


#Trading Hours Filter

market_open = "09:15"
market_close = "15:30"

spot = spot.between_time(market_open, market_close)
futures = futures.between_time(market_open, market_close)
options = options.between_time(market_open, market_close)


futures['open_interest'] = futures['open_interest'].clip(lower=0)


options['open_interest'] = options['open_interest'].clip(lower=0)


assert spot.index.equals(futures.index)
assert options.index.nunique() == spot.shape[0]


import os
os.makedirs("data/processed", exist_ok=True)

spot.to_csv("data/processed/nifty_spot_clean_5min.csv")
futures.to_csv("data/processed/nifty_futures_clean_5min.csv")
options.to_csv("data/processed/nifty_options_clean_5min.csv")
