==> Notebooks/03_feature_engineering.ipynb <==
import pandas as pd
import numpy as np


spot = pd.read_csv(
    "data/processed/nifty_spot_clean_5min.csv",
    parse_dates=['date'],
    index_col='date'
)

futures = pd.read_csv(
    "data/processed/nifty_futures_clean_5min.csv",
    parse_dates=['date'],
    index_col='date'
)

options = pd.read_csv(
    "data/processed/nifty_options_clean_5min.csv",
    parse_dates=['date'],
    index_col='date'
)


spot.shape, futures.shape, options.shape


spot['log_return'] = np.log(spot['close']).diff()
futures['log_return'] = np.log(futures['close']).diff()


spot['pct_return'] = spot['close'].pct_change()
futures['pct_return'] = futures['close'].pct_change()


spot['vol_30'] = spot['log_return'].rolling(30).std() * np.sqrt(30)
spot['vol_60'] = spot['log_return'].rolling(60).std() * np.sqrt(60)


for span in [9, 21, 50]:
    spot[f'ema_{span}'] = spot['close'].ewm(span=span, adjust=False).mean()
    futures[f'ema_{span}'] = futures['close'].ewm(span=span, adjust=False).mean()


spot['trend_9_21'] = spot['ema_9'] - spot['ema_21']
spot['trend_21_50'] = spot['ema_21'] - spot['ema_50']


 atm_options = options[options['strike'] == options['spot_close'].round(-2)]


from scipy.stats import norm

def bs_d1(S, K, T, r, sigma):
    return (np.log(S / K) + (r + 0.5 * sigma**2) * T) / (sigma * np.sqrt(T))

def delta(S, K, T, r, sigma, option_type):
    d1 = bs_d1(S, K, T, r, sigma)
    return norm.cdf(d1) if option_type == 'CE' else norm.cdf(d1) - 1

def gamma(S, K, T, r, sigma):
    d1 = bs_d1(S, K, T, r, sigma)
    return norm.pdf(d1) / (S * sigma * np.sqrt(T))


T = 1 / 252   # 1 trading day
r = 0.06


atm_options['delta'] = atm_options.apply(
    lambda x: delta(x['spot_close'], x['strike'], T, r, x['iv'], x['option_type']),
    axis=1
)

atm_options['gamma'] = atm_options.apply(
    lambda x: gamma(x['spot_close'], x['strike'], T, r, x['iv']),
    axis=1
)


pcr = (
    options
    .groupby([options.index, 'option_type'])['open_interest']
    .sum()
    .unstack()
)

pcr['pcr_oi'] = pcr['PE'] / pcr['CE']


spot = spot.merge(pcr[['pcr_oi']], left_index=True, right_index=True, how='left')


spot[['log_return', 'vol_30', 'ema_9', 'pcr_oi']].dropna().head()


spot.to_csv("data/processed/nifty_spot_features_5min.csv")

spot.columns
